Fixed indices i with not implemneted vertices

+++ source/TeleEngine.NET/Components/Vertices/SimpleShapeVertices/FloorComponent.cs
@@ -43,7 +43,7 @@ namespace TeleEngine.NET.Components.Vertices.SimpleShapeVertices
                    0.5f,  0, 0.5f,  0.0f, 1.0f,
                    -0.5f,  0f,  -0.5f,  0.0f, 1.0f,
             },
-            Indexes = new float[] 
+            Indexes = new uint[] 
             {
                0
             }
