+++ source/TeleEngine.NET/Components/Vertices/VertexModel.cs
@@ -4,11 +4,11 @@ namespace TeleEngine.NET.Components.Vertices
     public sealed class VertexModel
     {
         public Memory<float> Vertices { get; set; }
-        public Memory<float> Indexes { get; set; }
+        public Memory<uint> Indexes { get; set; }
 
         public VertexModel() { }
 
-        public VertexModel(Memory<float> vertices, Memory<float> indexes) 
+        public VertexModel(Memory<float> vertices, Memory<uint> indexes) 
         {
             Vertices = vertices;
             Indexes = indexes;
